requirements.txt
# Core Dependencies
neo4j>=5.14.0
PyYAML>=6.0.1
click>=8.1.7

# Database Connectors
mysql-connector-python>=8.2.0
psycopg2-binary>=2.9.9
pyodbc>=5.0.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Logging and Progress
tqdm>=4.66.1
colorlog>=6.8.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.1
flake8>=6.1.0
mypy>=1.7.1

# Utilities
python-dotenv>=1.0.0
